+++ src/commands/sdboot.sh
@@ -21,6 +21,8 @@
 #                           GLOBAL VARIABLES                          #
 #######################################################################
 
+SDBOOT_CMD_ADD=1
+SDBOOT_CMD_REMOVE=2
 SDBOOT_LOADER_CONF="${COMMON_ESP_PATH}/loader/loader.conf"
 SDBOOT_LOADER_ENTRIES_D="${COMMON_ESP_PATH}/loader/entries"
 SDBOOT_CONF_DEAFULT_KEY="default"
@@ -64,7 +66,7 @@ ${SDBOOT_LOADER_ENTRIES_D}/${conf_name}."
 }
 
 ###
-# Add sd-boot entry by creating loader conf file
+# Add uki sd-boot entry by creating loader conf file
 # ARGUMENTS:
 #   1 - uki path
 #   2 - efi dir
@@ -74,19 +76,13 @@ ${SDBOOT_LOADER_ENTRIES_D}/${conf_name}."
 # OUTPUTS:
 #   Debug info
 ###
-_sdboot_add_entry() {
+_sdboot_uki_add_entry() {
     uki="$1"
     efi_d="$2"
     arch="$3"
     kerver="$4"
     default="$5"
     common_install_uki_in_efi "$uki" "$efi_d"
-
-    case "$arch" in
-        aarch64) arch=aa64 ;;
-        x86_64)  arch=x64 ;;
-        # TODO: add more verification about possibles architecture
-    esac
 
     uki_file=$(basename "${uki}")
     uki_name=$(basename "${uki}" .efi)
@@ -105,7 +101,49 @@ EOF
 }
 
 ###
-# Remove the entry conf file
+# Add initrd sd-boot entry by creating loader conf file
+# ARGUMENTS:
+#   1 - initrd path
+#   2 - efi dir
+#   4 - kernel version
+#   5 - default option
+# OUTPUTS:
+#   Debug info
+###
+_sdboot_initrd_add_entry() {
+    initrd_path="$1"
+    efi_d="$2"
+    kerver="$3"
+    default="$4"
+    root_dev="$(common_get_dev_name /)"
+    root_uuid="$(common_get_dev_uuid "$root_dev")"
+    initrd_file=$(basename "${initrd_path}")
+
+    common_install_initrd_in_efi "${initrd_path}" "${kerver}"
+    common_get_machine_id
+    [ ! ${machine_id+x} ] && exit 2
+    esp_uname_d="${COMMON_ESP_PATH}/${machine_id}/${kerver}"
+    linux_file=$(find "${esp_uname_d}" -name "linux*")
+    linux_file=$(basename "${linux_file}")
+
+    cat > "${SDBOOT_LOADER_ENTRIES_D}/static-${machine_id}-${kerver}.conf" <<EOF
+title         Linux ${kerver}, static initrd ${initrd_file}
+sort-key      static-initrd
+version       ${kerver}
+machine-id    ${machine_id}
+options       root=UUID=${root_uuid} splash=silent mitigations=auto quiet \
+security=apparmor systemd.machine_id=${machine_id}
+linux         /${machine_id}/${kerver}/${linux_file}
+initrd        /${machine_id}/${kerver}/static-initrd
+EOF
+    echo_debug "initrd sdboot entry has been added."
+    if [ "${default}" = "1" ]; then
+        _sdboot_set_default "static-${machine_id}-${kerver}.conf"
+    fi
+}
+
+###
+# Remove the entry conf file of uki
 # ARGUMENTS:
 #   1 - uki path
 #   2 - kernel version
@@ -114,16 +152,38 @@ EOF
 # RETURN:
 #   exit 2 in error
 ###
-_sdboot_remove_entry() {
+_sdboot_uki_remove_entry() {
     uki="$1"
     kerver="$2"
 
     uki_file=$(basename "${uki}")
     uki_name=$(basename "${uki}" .efi)
-    conf_file="${SDBOOT_LOADER_ENTRIES_D}/${uki_name}_k${kerver}"
+    conf_file="${SDBOOT_LOADER_ENTRIES_D}/${uki_name}_k${kerver}.conf"
     if [ -f "${conf_file}" ]; then
-        rm "${SDBOOT_LOADER_ENTRIES_D}/${uki_name}_k${kerver}.conf"
+        rm "${conf_file}"
         echo_debug "UKI sdboot entry has been removed..."
+    else
+        echo_debug "No ${conf_file} to remove."
+    fi
+}
+
+###
+# Remove the entry conf file of initrd
+# ARGUMENTS:
+#   1 - kernel version
+# OUTPUTS:
+#   Debug info
+# RETURN:
+#   exit 2 in error
+###
+_sdboot_initrd_remove_entry() {
+    kerver="$1"
+    common_get_machine_id
+    [ ! ${machine_id+x} ] && exit 2
+    conf_file="${SDBOOT_LOADER_ENTRIES_D}/static-${machine_id}-${kerver}.conf"
+    if [ -f "${conf_file}" ]; then
+        rm "${conf_file}"
+        echo_debug "static initrd sdboot entry has been removed..."
     else
         echo_debug "No ${conf_file} to remove."
     fi
@@ -143,9 +203,9 @@ _sdboot_remove_entry() {
 _sdboot_usage() {
     usage_str="USAGE: $BIN sdboot [OPTIONS]
 OPTIONS:
-  --add:                Add entry
-  --remove:             Remove entry
+  --add | --remove:     Add / Remove sdboot entry (mandatory)
   -k|--kerver:          Kernel Version [Default: $KER_VER]
+  -i|--initrd:          Path to the initrd
   -u|--uki:             Path to the UKI name (should be end by .efi)
   -a|--arch:            Architecture to use [Default 'uname -m']
   -e|--efi:             efi directory [Default $COMMON_EFI_PATH]
@@ -153,17 +213,52 @@ OPTIONS:
   help:                 Print this helper
  
 INFO:
-    Create or remove a sdboot entry for the specified UKI.
-    If uki from path (--uki) point to a binary outside the boot partition, it
-    will try to install it into ${COMMON_ESP_PATH}/$efi_d.
-    If uki just mention an uki name file, it will search the binary from
-    '/usr/lib/modules/\$ker_ver/\$image'.
+  Create or remove a sdboot entry for the specified UKI or initrd.
+  If uki from path (--uki) point to a binary outside the boot partition, it \
+will try to install it into ${COMMON_ESP_PATH}/$efi_d.
+  If uki just mention an uki name file, it will search the binary from \
+'/usr/lib/modules/\$ker_ver/\$image'.
+  If the initrd provided isn't in the boot partition, it will copy it in \
+/boot 
  
 EXAMPLE:
   $BIN sdboot --add -k $(uname -r) -efi /EFI/opensuse -u uki-0.1.0.efi
-  $BIN sdboot --remove -k $(uname -r) -u uki-0.1.0.efi
-"
-    printf "%s\n" "$usage_str"
+  $BIN sdboot --remove -k $(uname -r) -u uki-0.1.0.efi"
+  printf "%s\n" "$usage_str"
+}
+
+###
+# Add or Remove UKI menue entry to grsdbootub2
+# ARGUMENTS:
+#   1 - commands [ADD/REMOVE]
+#   2 - kernel version
+#   3 - uki path
+#   4 - efi dir
+#   5 - arch
+#   6 - default option
+# RETURN:
+#   None
+###
+_sdboot_uki() {
+    # uki="$2"
+    # kerver="$3"
+    # efi_d="$4"
+    # arch="$5"
+    # default="$6"
+    if [ "$cmd" = "$SDBOOT_CMD_ADD" ]; then
+        _sdboot_uki_add_entry \
+            "${uki}" "${efi_d}" "${arch}" "${kerver}" "${default}"
+    else
+        _sdboot_uki_remove_entry "${uki}" "${kerver}"
+    fi
+}
+
+_sdboot_initrd() {
+    if [ "$cmd" = "$SDBOOT_CMD_ADD" ]; then
+        _sdboot_initrd_add_entry "${initrd}" "${efi_d}" "${kerver}" "${default}"
+    else
+        _sdboot_initrd_remove_entry "${kerver}"
+    fi
 }
 
 #######################################################################
@@ -203,32 +298,36 @@ sdboot_exec() {
     [ $# -lt 2 ] \
         && echo_error "Missing arguments"\
         && _extension_usage && exit 2
-    args=$(getopt -a -n extension -o u:,k:,a:,e:,D\
-        --long add,remove,kerver:,uki:,arch:,efi:,default -- "$@")
+    args=$(getopt -a -n extension -o u:,i:,k:,a:,e:,D\
+        --long add,remove,kerver:,initrd:,uki:,arch:,efi:,default -- "$@")
     eval set --"$args"
+    # Init some variables
+    kerver="$KER_VER"
+    arch=$(uname -m)
+    default=0
     while :
     do
         case "$1" in
-            --add)              cmd_add=1           ; shift 1 ;;
-            --remove)           cmd_remove=1        ; shift 1 ;;
-            -k | --kerver)      kerver="$2"         ; shift 2 ;;
-            -u | --uki)         uki="$2"            ; shift 2 ;;
-            -a | --arch)        arch="$2"           ; shift 2 ;;
-            -e | --efi)         efi_d="$2"          ; shift 2 ;;
-            -D | --default)     default=1           ; shift 1 ;;
-            --)                 shift               ; break   ;;
+            --add)              cmd_add=1       ; shift 1 ;;
+            --remove)           cmd_remove=1    ; shift 1 ;;
+            -k | --kerver)      kerver="$2"     ; shift 2 ;;
+            -i | --initrd)      initrd="$2"     ; shift 2 ;;
+            -u | --uki)         uki="$2"        ; shift 2 ;;
+            -a | --arch)        arch="$2"       ; shift 2 ;;
+            -e | --efi)         efi_d="$2"      ; shift 2 ;;
+            -D | --default)     default=1       ; shift 1 ;;
+            --)                 shift           ; break   ;;
             *) echo_warning "Unexpected option: $1"; _sdboot_usage   ;;
         esac
     done
-    if [ ! ${kerver+x} ]; then
-        kerver="$KER_VER"
-    fi
-    if [ ! ${uki+x} ]; then
-        echo_error "Missing uki name (--uki)"
+    case "$arch" in
+        aarch64) arch=aa64 ;;
+        x86_64)  arch=x64 ;;
+        # TODO: add more verification about possibles architecture
+    esac
+    if [ ! ${initrd+x} ] && [ ! ${uki+x} ]; then
+        echo_error "Missing uki or initrd name (--uki|--initrd)"
         exit 2
-    fi
-    if [ ! ${arch+x} ]; then
-        arch=$(uname -m)
     fi
     if [ ! ${efi_d+x} ]; then
         efi_d="$COMMON_EFI_PATH"
@@ -252,27 +351,31 @@ both!"
         exit 2
     elif [ ${cmd_add+x} ]; then
         _sdboot_install_bootctl
+        cmd=$SDBOOT_CMD_ADD
+    else
+        cmd=$SDBOOT_CMD_REMOVE
+    fi
+    # Check the mode
+    if [ ${initrd+x} ] && [ ${uki+x} ]; then
+        echo_error "Please choose between initrd or uki arguments. Not both!"
+        _sdboot_usage
+        exit 2
+    elif [ ! ${initrd+x} ] && [ ! ${uki+x} ]; then
+        echo_error "Missing initrd path OR uki path to add to the menu entry"
+        _sdboot_usage
+        exit 2
+    elif [ ${uki+x} ]; then
         if [ ! -f "${uki}" ]; then
             uki_file=$(basename "${uki}")
             uki="/usr/lib/modules/${kerver}/${uki_file}"
         fi
-        _sdboot_add_entry "${uki}" "${efi_d}" "${arch}" "${kerver}" "${default}"
-        # err=$(sdbootutil \
-        #     --arch="$arch" \
-        #     --image="$uki" \
-        #     add-uki "$kerver" 2>&1)
-        # ret=$?
-    else
-        _sdboot_remove_entry "${uki}" "${kerver}"
-        # err=$(sdbootutil \
-        #     --arch="$arch" \
-        #     --image="$uki" \
-        #     remove-uki "$kerver" \
-        #     2>&1)
-        # ret=$?
+        _sdboot_uki
+    elif [ ${initrd+x} ]; then
+        if [ ! -f "${initrd}" ]; then
+            initrd_file=$(basename "${initrd}")
+            initrd="/usr/share/initrd/${initrd_file}"
+        fi
+        _sdboot_initrd
     fi
-    # if [ $ret -ne 0 ]; then
-    #     echo_error "sdbootutil : '$err'"
-    #     exit 2
-    # fi
+
 }